requirements.txt
fastapi>=0.135.1
uvicorn>=0.41.0
anthropic>=0.84.0
python-dotenv>=1.2.2
httpx>=0.28.1
pydantic>=2.12.5
schedule>=1.2.2
rich>=14.3.3
sqlalchemy>=2.0.48
psycopg2-binary>=2.9.11
apscheduler>=3.11.2
stripe>=12.0.0
stripe>=12.0.0
